rag-python/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
requests
python-docx
tika
